// src/utils/enums/requests-enum.ts
export enum RequestsType {
  CREATED = "CREATED",
  CANCELED = "CANCELED",
  CANCELED_REQUESTED = "CANCELED_REQUESTED",
  CANCEL_REJECTED = "CANCEL_REJECTED",
  EDITED = "EDITED",
  REJECTED = "REJECTED",
  APPROVED = "APPROVED",
  PEOPLE_REQUESTED = "PEOPLE_REQUESTED",
  PEOPLE_SENT = "PEOPLE_SENT",
  PAYMENT_REQUESTED = "PAYMENT_REQUESTED",
  PAYMENT_SENT = "PAYMENT_SENT",
  PAYMENT_APPROVED = "PAYMENT_APPROVED",
  PAYMENT_REJECTED = "PAYMENT_REJECTED",
}

export enum ProfessorRequestsType {
  DOCUMENT_REQUESTED = "DOCUMENT_REQUESTED",
  DOCUMENT_APPROVED = "DOCUMENT_APPROVED",
  DOCUMENT_REJECTED = "DOCUMENT_REJECTED",
}